+++ lib/kafka_consumer.ex
@@ -1,75 +1,56 @@
 defmodule KafkaConsumer do
-  use GenServer
-  alias KafkaConsumer.Utils
+  use Application
+  import Supervisor.Spec, only: [worker: 3]
+  alias KafkaConsumer.{Server, Utils}
 
   defmodule Settings do
-    defstruct topic: "",
-     partition: 0,
-     handler: nil,
-     handler_pool: nil
+    defstruct topic: nil, partition: 0, handler: nil, handler_pool: nil
   end
 
-  @type settings :: %Settings{topic: String.t,
-   partition: number,
-   handler: term(),
-   handler_pool: atom
- }
+  @type settings :: %Settings{
+    topic: String.t | nil,
+    partition: non_neg_integer(),
+    handler: atom(),
+    handler_pool: atom()
+  }
 
-  def start_link(%{topic: topic, partition: partition} = settings) do
-    consumer_name = [topic, partition] |> Enum.join("$")
-    GenServer.start_link(__MODULE__, settings, name: via_tuple(consumer_name))
+  def start(_type, _args) do
+    children = event_handlers_spec(event_handlers)
+
+    opts = [strategy: :one_for_one, name: KafkaConsumer.Supervisor]
+    Supervisor.start_link(children, opts)
   end
 
-  ### GenServer Callbacks
-
-  def init(%{topic: topic, partition: partition} = settings) do
-    worker_name = [topic, partition, "stream"] |> Enum.join("$") |> String.to_atom
-    send self(), {:consume, Map.put(settings, :worker_name, worker_name)}
-
-    {:ok, %{topic: topic, partition: partition, worker_name: worker_name}}
+  defp event_handlers do
+    Application.get_env(:kafka_consumer, :event_handlers)
   end
 
-  def handle_info({:consume, settings}, state) do
-    spawn_link(fn -> consume(settings, self()) end)
-    {:noreply, state}
+  defp event_handlers_spec(event_handlers) do
+    childs = Enum.map(event_handlers, fn {topic, partition, handler, handler_pool,
+      size, max_overflow} ->
+        pool = poolboy_spec(handler_pool, handler, size, max_overflow)
+        worker = worker(Server,
+          [%KafkaConsumer.Settings{topic: topic,
+              partition: partition,
+              handler: handler,
+              handler_pool: handler_pool}],
+          [id: [topic, partition] |> Enum.join("$") |> String.to_atom])
+        [pool, worker]
+    end)
+
+    childs
+    |> List.flatten
+    |> Enum.uniq
   end
 
-  def handle_info(:topic_not_found, state) do
-    {:stop, :topic_not_found, state}
-  end
+  defp poolboy_spec(name, handler, size, max_overflow) do
+    poolboy_config = [
+      {:name, {:local, name}},
+      {:worker_module, handler},
+      {:size, size},
+      {:max_overflow, max_overflow}
+    ]
 
-  def terminate(:topic_not_found, _state) do
-    :ok
-  end
-
-  def terminate(_reason, %{worker_name: worker_name} = _state) do
-    Utils.stop_stream(worker_name)
-    :ok
-  end
-
-  ### Internal Functions
-
-  @doc """
-    Consume messages from topic or stop consumer if topic not found
-  """
-  @spec consume(Map.t, pid) :: atom
-  def consume(%{topic: topic, partition: partition, worker_name: worker_name,
-      handler: handler, handler_pool: handler_pool}, consumer) do
-    if Utils.topic_exists?(topic, partition) do
-      Utils.prepare_stream(worker_name)
-      for message <- KafkaEx.stream(topic, partition, worker_name: worker_name) do
-        :poolboy.transaction(handler_pool, fn(pid) ->
-          handler.handle_event(pid, {topic, partition, message})
-        end)
-      end
-    else
-      send consumer, :topic_not_found
-    end
-    :ok
-  end
-
-  @spec via_tuple(String.t) :: {atom, atom, {atom, atom, {atom, String.t}}}
-  defp via_tuple(worker_name) do
-    {:via, :gproc, {:n, :l, {:consumer, worker_name}}}
+    :poolboy.child_spec(name, poolboy_config, [])
   end
 end
